# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Swagger Api dynamic host server"
        }
    ],
    "paths": {
...
        "/predictions": {
            "get": {
                "tags": [
                    "Prediction"
                ],
                "summary": "getPrediction",
                "description": "get prediction",
                "operationId": "getPrediction",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorting",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "latest",
                                "betAmount"
                            ]
                        }
                    },
                    {
                        "name": "eventId",
                        "in": "query",
                        "description": "user id to view all his forecasts"
                    },
                    {
                        "name": "dateFrom",
                        "in": "query",
                        "description": "Only predictions added after this date"
                    },
                    {
                        "name": "dateTo",
                        "in": "query",
                        "description": "Only predictions added before this date"
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "The number of predictions issued at a time per page",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number"
                    },
                    {
                        "name": "isApproved",
                        "in": "query",
                        "description": "show predictions by flag isApproved",
                        "schema": {
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    },
                    {
                        "name": "takeForModeration",
                        "in": "query",
                        "description": "If true, all returned predictions moderated",
                        "schema": {
                            "enum": [
                                "true",
                                "false"
                            ]
                        }
                    },
                    {
                        "name": "moderatorId",
                        "in": "query",
                        "description": "returns predictions that were taken by a particular moderator"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "success"
                                            ]
                                        },
                                        "prediction": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "request error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "fail"
                                            ]
                                        },
                                        "errorCode": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Prediction"
                ],
                "summary": "createPrediction",
                "description": "create prediction",
                "operationId": "createPrediction",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "description": "prediction description",
                                        "type": "string"
                                    },
                                    "betPercentage": {
                                        "description": "percentage of money",
                                        "type": "integer"
                                    },
                                    "events": {
                                        "description": "events list",
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "success"
                                            ]
                                        },
                                        "prediction": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "request error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "fail"
                                            ]
                                        },
                                        "errorCode": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/predictions/{id}": {
            "get": {
                "tags": [
                    "Prediction"
                ],
                "summary": "getPrediction",
                "description": "get prediction by id",
                "operationId": "getPrediction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "prediction id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "success"
                                            ]
                                        },
                                        "prediction": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "request error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "fail"
                                            ]
                                        },
                                        "errorCode": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "bearer",
                "scheme": "bearer"
            }
        }
    },
...
}